Support dnsu2t redsocks proxy config
Users may specify dnsu2t config by including a `dns` field
in their PATCH body:
```
{
  network: {
    dns: '1.1.1.1:53',
  }
}
```

If `dns` is a string, IP and PORT are optional but should be in
the format [IP][:PORT] if present. For any fields that are empty,
a default IP and PORT will be used.

`dns` may also be a boolean. If true, defaults will be configured.
If false, the dns configuration will be removed.

If `proxy` is patched to empty, `dns` will be removed regardless of its
current or input configs, as `dns` depends on an active redsocks
proxy to function.

Change-type: minor
Signed-off-by: Christina Ying Wang <christina@balena.io>